new(vx-demo): convert Brush to codesandbox
