+++ packages/vx-demo/src/docs-v2/examples/vx-brush/AreaChart.tsx
@@ -1,0 +1,92 @@
+import React from 'react';
+import { Group } from '@vx/group';
+import { AreaClosed } from '@vx/shape';
+import { ScaleType } from '@vx/shape/lib/types';
+import { AxisLeft, AxisBottom } from '@vx/axis';
+import { LinearGradient } from '@vx/gradient';
+import { curveMonotoneX } from '@vx/curve';
+import { AppleStock } from '@vx/mock-data/lib/mocks/appleStock';
+
+// Initialize some variables
+const axisColor = '#fff';
+const axisBottomTickLabelProps = {
+  textAnchor: 'middle' as const,
+  fontFamily: 'Arial',
+  fontSize: 10,
+  fill: axisColor,
+};
+const axisLeftTickLabelProps = {
+  dx: '-0.25em',
+  dy: '0.25em',
+  fontFamily: 'Arial',
+  fontSize: 10,
+  textAnchor: 'end' as const,
+  fill: axisColor,
+};
+
+// accessors
+const getDate = (d: AppleStock) => new Date(d.date);
+const getStockValue = (d: AppleStock) => d.close;
+
+export default function AreaChart({
+  data,
+  width,
+  yMax,
+  margin,
+  xScale,
+  yScale,
+  hideBottomAxis = false,
+  hideLeftAxis = false,
+  top,
+  left,
+  children,
+}: {
+  data: AppleStock[];
+  xScale: ScaleType;
+  yScale: ScaleType;
+  width: number;
+  yMax: number;
+  margin: { top: number; right: number; bottom: number; left: number };
+  hideBottomAxis?: boolean;
+  hideLeftAxis?: boolean;
+  top?: number;
+  left?: number;
+  children?: React.ReactNode;
+}) {
+  return (
+    <Group left={left || margin.left} top={top || margin.top}>
+      <LinearGradient id="gradient" from="#fff" fromOpacity={1} to="#fff" toOpacity={0.2} />
+
+      {!hideBottomAxis && (
+        <AxisBottom<Date>
+          top={yMax}
+          scale={xScale}
+          numTicks={width > 520 ? 10 : 5}
+          stroke={axisColor}
+          tickStroke={axisColor}
+          tickLabelProps={() => axisBottomTickLabelProps}
+        />
+      )}
+      {!hideLeftAxis && (
+        <AxisLeft<number>
+          scale={yScale}
+          numTicks={5}
+          stroke={axisColor}
+          tickStroke={axisColor}
+          tickLabelProps={() => axisLeftTickLabelProps}
+        />
+      )}
+      <AreaClosed<AppleStock>
+        data={data}
+        x={d => xScale(getDate(d)) || 0}
+        y={d => yScale(getStockValue(d)) || 0}
+        yScale={yScale}
+        strokeWidth={1}
+        stroke="url(#gradient)"
+        fill="url(#gradient)"
+        curve={curveMonotoneX}
+      />
+      {children}
+    </Group>
+  );
+}
