+++ packages/vx-demo/src/docs-v2/examples/vx-brush/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "@vx/demo-brush",
+  "description": "Standalone vx brush demo.",
+  "main": "index.tsx",
+  "dependencies": {
+    "@babel/runtime": "^7.8.4",
+    "@types/react": "^16",
+    "@types/react-dom": "^16",
+    "@vx/axis": "latest",
+    "@vx/brush": "latest",
+    "@vx/curve": "latest",
+    "@vx/gradient": "latest",
+    "@vx/group": "latest",
+    "@vx/mock-data": "latest",
+    "@vx/pattern": "latest",
+    "@vx/responsive": "latest",
+    "@vx/scale": "latest",
+    "@vx/shape": "latest",
+    "d3-array": "^2.4.0",
+    "react": "^16.8",
+    "react-dom": "^16.8",
+    "react-scripts-ts": "3.1.0",
+    "typescript": "^3"
+  },
+  "keywords": [
+    "visualization",
+    "d3",
+    "react",
+    "vx",
+    "brush",
+    "interaction"
+  ]
+}
