+++ packages/vx-demo/src/pages/Brush.tsx
@@ -1,21 +1,20 @@
 import React from 'react';
 import Show from '../components/Show';
-import Brush from '../components/tiles/Brush';
-import BrushSource from '!!raw-loader!../components/tiles/Brush';
+import Brush from '../docs-v2/examples/vx-brush/Example';
+import BrushSource from '!!raw-loader!../docs-v2/examples/vx-brush/Example';
 
-export default () => {
-  return (
-    <Show
-      component={Brush}
-      title="Brush"
-      margin={{
-        top: 50,
-        left: 50,
-        right: 20,
-        bottom: 50,
-      }}
-    >
-      {BrushSource}
-    </Show>
-  );
-};
+export default () => (
+  <Show
+    component={Brush}
+    title="Brush"
+    margin={{
+      top: 50,
+      left: 50,
+      right: 20,
+      bottom: 50,
+    }}
+    codeSandboxDirectoryName="vx-brush"
+  >
+    {BrushSource}
+  </Show>
+);
